--- expressapp.js
@@ -1,15 +1,0 @@
-const express = require('express')
-const app = express()
-const port = 3000
-
-app.get('/', (req, res) => {
-  res.send('This is Home Page!')
-})
-
-app.get('/about', (req, res) => {
-    res.send('This is About Page!')
-})
-
-app.listen(port, () => {
-  console.log(`Example app listening on port ${port}`)
-})
