--- file.txt
@@ -1,1 +1,0 @@
-THIS IS NEW
