--- second.js
@@ -1,7 +1,0 @@
-    pavan = {
-        name:"pavankuamr",
-        age: 22
-    }
-
-module.exports = pavan         
-
